Build overrides in Copr and report status of the builds
It also changes how the builds are tracked, we are keeping directly the
`build_proxy` response from Copr. This is basically a dictionary (which
Copr calls `Munch`) of the API response.

Since we want to store this for component's default build and also for
overrides it is a nice structure representing the build.