Consistently generate pkg nevra even if the last commit is release
When upstream's last commit is the release commit we get release == 1 and
RGO generated the package nevra without date and hash, it could look
like a normal fedora release: libdnf-0.66.0-1.fc35.x86_64.rpm.

If the pkg was just released to Fedora both our build and fedora
build would have the same nevra. This means that even if the package is
available it doesn't get upgraded which could be a problem because the
fedora build can contain downstream only patches and thus behave
differently than our RGO build.

This is causing fails in our CI because we test for unpatched behavior
but since libdnf doesn't get upgraded we get a patched build from fedora.